requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
pandas>=2.0.0
bcrypt>=4.0.1
pyotp>=2.9.0
PyJWT>=2.8.0
reportlab>=4.0.8
matplotlib>=3.8.0